+++ extension/sessionsifu@local/ui/sessionItemButtons.js
@@ -227,7 +227,6 @@ class SessionItemButtons extends GObject.Object {
 
     _onClickRestore(button, event) {
         Autoclose.autocloseObject.sessionClosedByUser = false;
-        RestoreSession.restoreSessionObject.restoringApps = new Map();
         // Using _restoredApps to hold restored apps so we create new instance every time for now
         const _restoreSession = new RestoreSession.RestoreSession();
         _restoreSession.restoreSession(this.sessionItem._filename);